+++ docs/changelog.md
@@ -1,11 +1,30 @@
 # Changelog
 
-## v0.0.1 - Date
+## [v0.1.0] - 11 July 2024
 
-**Improvement**:
 
--   TBD
+**What's Changed**
 
-**New Features**:
+- Add support for token streaming and custom CORS by [@Blaizzy](https://github.com/Blaizzy)
+- Add support for Parallel calls by [@Blaizzy](https://github.com/Blaizzy)
+- Add Parallel calls usage  by [@Blaizzy](https://github.com/Blaizzy)
 
--   TBD
+**Fixes :**
+
+- Cross origin Support [#2](https://github.com/Blaizzy/fastmlx/issues/2)
+- Max tokens not overriding [#5](https://github.com/Blaizzy/fastmlx/issues/5)
+
+## [v0.0.1] - 09 July 2024
+
+
+**What's Changed**
+
+- Setup FastMLX by [@Blaizzy](https://github.com/Blaizzy)
+- Add support for VLMs by [@Blaizzy](https://github.com/Blaizzy)
+- Add support for LMs by by [@Blaizzy](https://github.com/Blaizzy)
+
+**New Contributors**
+
+- [@Blaizzy](https://github.com/Blaizzy) made their first contribution in [https://github.com/Blaizzy/fastmlx/pull/1](https://github.com/Blaizzy/fastmlx/pull/1)
+
+